Exclude WinRTInteropCallbacks from non-WinRT builds of CoreLib
I'm mostly doing this because I find it offensive to have WinRTInterop callbacks in Win32 builds - I'm not blocked on this in any way.

This was supposed to be an easy change, but exposed an annoyance in how Win32 builds got hooked up into the build. Apparently, we were building WinRT version of S.P.Interop against the Win32 version of S.P.CoreLib because both Win32 and UWP versions of class libraries were stomping over each other's build outputs. The change in open.targets is supposed to address that.

[tfs-changeset: 1664306]
